# autoevolve-ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Company Cockpit",
    "version": "0.1.0"
  },
  "paths": {
...
    "/company/sales/contacts": {
      "get": {
        "tags": [
          "contacts"
        ],
        "summary": "Contacts",
        "operationId": "contacts_company_sales_contacts_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Page Size"
            }
          },
          {
            "name": "stage",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "new",
                    "enriched",
                    "qualified",
                    "draft_ready",
                    "approved",
                    "contacted",
                    "scheduled",
                    "replied",
                    "won",
                    "lost",
                    "suppressed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stage"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "recent",
                "oldest",
                "name",
                "score",
                "company"
              ],
              "type": "string",
              "default": "recent",
              "title": "Sort"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactsPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/company/sales/contacts/{contact_id}": {
      "get": {
        "tags": [
          "contacts"
        ],
        "summary": "Contact",
        "operationId": "contact_company_sales_contacts__contact_id__get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "contact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompanyProfile": {
        "properties": {
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          },
          "fetched_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetched At"
          }
        },
        "type": "object",
        "required": [
          "domain",
          "provider",
          "status",
          "summary"
        ],
        "title": "CompanyProfile"
      },
...
      "Contact": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "job_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Title"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "company_domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Domain"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "source_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Detail"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "verification_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Status"
          },
          "lead_score": {
            "type": "integer",
            "title": "Lead Score"
          },
          "stage": {
            "type": "string",
            "enum": [
              "new",
              "enriched",
              "qualified",
              "draft_ready",
              "approved",
              "contacted",
              "scheduled",
              "replied",
              "won",
              "lost",
              "suppressed"
            ],
            "title": "Stage"
          },
          "suppression_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suppression Reason"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "last_touch_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Touch At"
          },
          "contact_profile": {
            "$ref": "#/components/schemas/ContactProfile"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "lead_score",
          "stage",
          "created_at",
          "updated_at",
          "contact_profile"
        ],
        "title": "Contact"
      },
      "ContactDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "job_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Title"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "company_domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Domain"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "source_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Detail"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "verification_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Status"
          },
          "lead_score": {
            "type": "integer",
            "title": "Lead Score"
          },
          "stage": {
            "type": "string",
            "enum": [
              "new",
              "enriched",
              "qualified",
              "draft_ready",
              "approved",
              "contacted",
              "scheduled",
              "replied",
              "won",
              "lost",
              "suppressed"
            ],
            "title": "Stage"
          },
          "suppression_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suppression Reason"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "last_touch_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Touch At"
          },
          "contact_profile": {
            "$ref": "#/components/schemas/ContactProfile"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "company_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompanyProfile"
              },
              {
                "type": "null"
              }
            ]
          },
          "drafts": {
            "items": {
              "$ref": "#/components/schemas/DraftSummary"
            },
            "type": "array",
            "title": "Drafts"
          },
          "interactions": {
            "items": {
              "$ref": "#/components/schemas/InteractionSummary"
            },
            "type": "array",
            "title": "Interactions"
          },
          "recent_events": {
            "items": {
              "$ref": "#/components/schemas/EventSummary"
            },
            "type": "array",
            "title": "Recent Events"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "lead_score",
          "stage",
          "created_at",
          "updated_at",
          "contact_profile",
          "metadata",
          "drafts",
          "interactions",
          "recent_events"
        ],
        "title": "ContactDetail"
      },
      "ContactMetrics": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "ready": {
            "type": "integer",
            "title": "Ready"
          },
          "verified": {
            "type": "integer",
            "title": "Verified"
          },
          "suppressed": {
            "type": "integer",
            "title": "Suppressed"
          },
          "by_stage": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Stage"
          }
        },
        "type": "object",
        "required": [
          "total",
          "ready",
          "verified",
          "suppressed",
          "by_stage"
        ],
        "title": "ContactMetrics"
      },
      "ContactProfile": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "linkedin_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkedin Url"
          },
          "email_ready": {
            "type": "boolean",
            "title": "Email Ready"
          },
          "ready_for_outreach": {
            "type": "boolean",
            "title": "Ready For Outreach"
          },
          "verification_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Status"
          }
        },
        "type": "object",
        "required": [
          "status",
          "email_ready",
          "ready_for_outreach"
        ],
        "title": "ContactProfile"
      },
      "ContactsPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Contact"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "metrics": {
            "$ref": "#/components/schemas/ContactMetrics"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "metrics"
        ],
        "title": "ContactsPage"
      },
...
      "DraftSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "subject",
          "status",
          "created_at"
        ],
        "title": "DraftSummary"
      },
...
      "EventSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "event": {
            "type": "string",
            "title": "Event"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event",
          "created_at"
        ],
        "title": "EventSummary"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InteractionSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "direction": {
            "type": "string",
            "title": "Direction"
          },
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "direction",
          "channel",
          "kind",
          "created_at"
        ],
        "title": "InteractionSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}